requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
fastapi==0.113.0
greenlet==3.0.3
idna==3.8
pydantic==2.9.0
pydantic_core==2.23.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
python-dotenv==1.0.1
psycopg2
jinja2
python-multipart
requests
aiohttp